fix: contacts page
