+++ app/shared/components/column-guides/ColumnGuides.tsx
@@ -2,7 +2,7 @@
 
 import { Box, useMediaQuery, useTheme } from '@mui/material';
 
-import { containerStyle, getLineSideStyle, gridContainerStyle } from './ColumnGuides.style';
+import { styles } from './ColumnGuides.style';
 
 interface ColumnGuidesProps {
   lineColor?: string;
@@ -70,14 +70,14 @@ export const ColumnGuides = ({ lineColor = 'rgba(237, 232, 223, 1)' }: ColumnGui
   });
 
   return (
-    <Box sx={containerStyle}>
-      <Box aria-hidden sx={gridContainerStyle(layout, gap, paddingX)}>
+    <Box sx={styles.containerStyle}>
+      <Box aria-hidden sx={styles.gridContainerStyle(layout, gap, paddingX)}>
         {Object.entries(grouped).map(([colStr, aligns]) => {
           const col = Number(colStr);
           return (
             <Box key={`col-${col}`} sx={{ gridColumn: col, position: 'relative' }}>
-              {aligns.includes('start') && <Box sx={getLineSideStyle('left', lineColor)} />}
-              {aligns.includes('end') && <Box sx={getLineSideStyle('right', lineColor)} />}
+              {aligns.includes('start') && <Box sx={styles.getLineSideStyle('left', lineColor)} />}
+              {aligns.includes('end') && <Box sx={styles.getLineSideStyle('right', lineColor)} />}
             </Box>
           );
         })}
